Fix floating-point in arm64, add tests

+++ testcases/test_floating_ret_val/test_func_no_libc.c
@@ -1,6 +1,6 @@
 float func_1(
-    float dividend,
-    float divisor
+    int dividend,
+    int divisor
 )
 {    
     float res = dividend / divisor;
